+++ example/data-persistence-tf/variables.tf
@@ -96,7 +96,9 @@ variable "subnets_tag_name" {
 variable "cloudwatch_log_retention_periods" {
   type = map(number)
   description = "number of days logs will be retained for the respective cloudwatch log group, in the form of <cloudwatch_log_group_name>_log_retention"
-  default = {}
+  default = {
+    postgres-db-migration = 7, # data-persistence module
+  }
 }
 
 variable "default_log_retention_days" {