+++ lambdas/data-migration1/variables.tf
@@ -68,5 +68,5 @@ variable "cloudwatch_log_retention_periods" {
 variable "default_log_retention_days" {
   type = number
   default = 30
-  description = "default cloudwatch log retention periods"
+  description = "default value that user chooses for their log retention periods"
 }