+++ lambdas/postgres-migration-async-operation/variables.tf
@@ -49,6 +49,7 @@ variable "permissions_boundary_arn" {
 variable "prefix" {
   type = string
 }
+
 variable "rds_connection_timing_configuration" {
   description = "Cumulus rds connection timeout retry timing object -- these values map to knex.js's internal use of  https://github.com/vincit/tarn.js/ for connection acquisition"
   type = map(number)
@@ -88,12 +89,12 @@ variable "vpc_id" {
 
 variable "cloudwatch_log_retention_periods" {
   type = map(number)
-  description = "retention periods for the respective cloudwatch log group, these values will be used instead of default retention days"
+  description = "Optional retention periods for the respective cloudwatch log group, these values will be used instead of default retention days"
   default = {}
 }
 
 variable "default_log_retention_days" {
   type = number
   default = 30
-  description = "default cloudwatch log retention periods"
+  description = "Optional default value that user chooses for their log retention periods"
 }