+++ tf-modules/distribution/variables.tf
@@ -24,7 +24,7 @@ variable "tea_internal_api_endpoint" {
 
 variable "cloudwatch_log_retention_periods" {
   type = map(number)
-  description = "number of days logs will be retained for the respective cloudwatch log group, in the form of <cloudwatch_log_group_name>_log_retention"
+  description = "retention periods for the respective cloudwatch log group, these values will be used instead of default retention days"
   default = {}
 }