+++ tf-modules/ingest/variables.tf
@@ -4,18 +4,6 @@ variable "buckets" {
   default = {}
 }
 
-variable "cloudwatch_log_retention_periods" {
-  type = map(number)
-  description = "number of days logs will be retained for the respective cloudwatch log group, in the form of <cloudwatch_log_group_name>_log_retention"
-  default = {}
-}
-
-variable "default_log_retention_days" {
-  type = number
-  default = 30
-  description = "default value that user chooses for their log retention periods"
-}
-
 variable "cmr_oauth_provider" {
   type = string
 }
@@ -132,7 +120,6 @@ variable "lzards_provider" {
   default     = ""
 }
 
-
 variable "permissions_boundary_arn" {
   type = string
 }
@@ -184,4 +171,16 @@ variable "throttled_queues" {
 variable "vpc_id" {
   type    = string
   default = null
+}
+
+variable "cloudwatch_log_retention_periods" {
+  type = map(number)
+  description = "Optional retention periods for the respective cloudwatch log group, these values will be used instead of default retention days"
+  default = {}
+}
+
+variable "default_log_retention_days" {
+  type = number
+  default = 30
+  description = "Optional default value that user chooses for their log retention periods"
 }