+++ tf-modules/tea-map-cache/variables.tf
@@ -34,12 +34,12 @@ variable "deploy_to_ngap" {
 
 variable "cloudwatch_log_retention_periods" {
   type = map(number)
-  description = "number of days logs will be retained for the respective cloudwatch log group, in the form of <cloudwatch_log_group_name>_log_retention"
+  description = "Optional retention periods for the respective cloudwatch log group, these values will be used instead of default retention days"
   default = {}
 }
 
 variable "default_log_retention_days" {
   type = number
   default = 30
-  description = "default value that user chooses for their log retention periods"
+  description = "Optional default value that user chooses for their log retention periods"
 }